[#64] API 서비스 중복 코드 정리 및 S3 업로드 헬퍼 추출
