+++ Codive/Features/Closet/Data/ClothAPIService+Helpers.swift
@@ -7,25 +7,10 @@
 
 import Foundation
 import CodiveAPI
-import CryptoKit
 
-// MARK: - Private Helpers
+// MARK: - Season Mapping Helpers
 
 extension ClothAPIService {
-
-    func calculateMD5(from data: Data) -> String {
-        let digest = Insecure.MD5.hash(data: data)
-        return Data(digest).base64EncodedString()
-    }
-
-    func extractFinalUrl(from presignedUrl: String) -> String {
-        guard let url = URL(string: presignedUrl),
-              var components = URLComponents(url: url, resolvingAgainstBaseURL: false) else {
-            return presignedUrl
-        }
-        components.query = nil
-        return components.string ?? presignedUrl
-    }
 
     func mapSeasonToCreatePayload(_ season: Season) -> Components.Schemas.ClothCreateRequest.seasonsPayloadPayload {
         switch season {
