+++ Codive/Features/Feed/Domain/UseCases/DeleteHistoryUseCase.swift
@@ -2,7 +2,7 @@
 //  DeleteHistoryUseCase.swift
 //  Codive
 //
-//  Created by Claude on 2026-02-12.
+//  Created by 황상환 on 2026-02-12.
 //
 
 import Foundation
