+++ Codive/Features/LookBook/Data/Services/LookBookAPIService.swift
@@ -8,7 +8,6 @@
 import Foundation
 import CodiveAPI
 import OpenAPIRuntime
-import CryptoKit
 
 // MARK: - HomeCategoryAPIService Protocol
